Fix bug deleting non existing entries

+++ cg/store/api/delete.py
@@ -15,22 +15,17 @@ class DeleteDataHandler(BaseHandler):
             flowcell.flush()
             self.commit()
 
-    def delete_relationships_case(self, case: Family) -> None:
-        """Delete association entries between a case and its samples."""
-        for case_sample in case.links:
-            case_sample.delete()
-        self.commit()
-
     def delete_relationships_sample(self, sample: Sample):
-        """Delete association between all cases and the provided sample."""
-        for case_sample in sample.links:
-            case_sample.delete()
-        self.commit()
+        """Delete relationships between all cases and the provided sample."""
+        if sample and sample.links:
+            for case_sample in sample.links:
+                case_sample.delete()
+            self.commit()
 
-    def delete_cases_without_samples(self, case_ids: List[str]) -> List[str]:
+    def delete_cases_without_samples(self, case_ids: List[str]) -> None:
         """Delete cases without samples."""
         for case_id in case_ids:
             case: Family = self.Family.query.filter(Family.internal_id == case_id).first()
-            if not case.links:
+            if case and not case.links:
                 case.delete()
         self.commit()
